-Camera still has billy as a variable name
-Border layer is sus and can be inproved
-can shoot and aim in pause menu
-meth is not impimented
-can shoot below the floor
-gun is too big enemy cant get killed

+++ Assets/Scripts/spawn_enemy.cs
@@ -14,7 +14,7 @@ public class spawn_enemy : MonoBehaviour
     public CompositeCollider2D floor_collider;
     public Camera camera;
     [SerializeField] private bool spawnEnemy;
-    [SerializeField] private int spawnSpeed;
+    [SerializeField] private float spawnSpeed;
